+++ app/src/main/java/com/example/weatherapp/data/network/entities/polution/Item.kt
@@ -2,15 +2,15 @@ package com.example.weatherapp.data.network.entities.polution
 
 import com.example.weatherapp.domain.model.PolutionDetails
 
-data class Item (
+data class Item(
     val components: Components,
     val dt: Int,
     val main: Main
 ) {
-    fun map() : PolutionDetails{
+    fun map(): PolutionDetails {
         return PolutionDetails(
             components.co,
-            components.no,
+            components.no.toDouble(),
             components.no2,
             components.o3,
             components.so2,
